+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.29.2] - 2025-05-15
 ### Added:
 - Tree Render: `render_tree` to support drag-and-drop to shift nodes.
+### Fixed:
+- Tree Render: Make tkinter module import fail-safe.
 
 ## [0.29.1] - 2025-05-11
 ### Added:
@@ -792,7 +796,8 @@ ignore null attribute columns.
 - Utility Iterator: Tree traversal methods.
 - Workflow To Do App: Tree use case with to-do list implementation.
 
-[Unreleased]: https://github.com/kayjan/bigtree/compare/0.29.1...HEAD
+[Unreleased]: https://github.com/kayjan/bigtree/compare/0.29.2...HEAD
+[0.29.2]: https://github.com/kayjan/bigtree/compare/0.29.1...0.29.2
 [0.29.1]: https://github.com/kayjan/bigtree/compare/0.29.0...0.29.1
 [0.29.0]: https://github.com/kayjan/bigtree/compare/0.28.0...0.29.0
 [0.28.0]: https://github.com/kayjan/bigtree/compare/0.27.0...0.28.0